Rename CompositionHeader to CompositionFormHeader
